requirements.txt
Django==1.9.5
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-filter==0.13.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
Markdown==2.6.6
markdown2==2.3.6
Pillow==3.1.0
PyJWT==1.4.0
six==1.11.0
typing==3.6.6
